% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunner.munzel.perutation.test.R
\name{brunner.munzel.permutation.test}
\alias{brunner.munzel.permutation.test}
\alias{brunner.munzel.permutation.test.default}
\alias{brunner.munzel.permutation.test.formula}
\title{permuted Brunner-Munzel test}
\usage{
brunner.munzel.permutation.test(x, ...)

\method{brunner.munzel.permutation.test}{default}(x, y,
  alternative = c("two.sided", "greater", "less"), ...)

\method{brunner.munzel.permutation.test}{formula}(formula, data, subset,
  na.action, ...)
}
\arguments{
\item{x}{the numeric vector of data values from the sample 1.}

\item{\dots}{further arguments to be passed to or from methods
(This argument is for only formula).}

\item{y}{the numeric vector of data values from the sample 2.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. User can specify just the initial letter.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs}
is a numeric variable giving the data values and \code{rhs} a factor
with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to
\code{getOption("na.action")}.}
}
\value{
A list containing the following components:
 \item{method}{the characters ``Brunner-Munzel Permutation Test''}
 \item{p.value}{the \eqn{p}-value of the test.}
 \item{data.name}{a character string giving the name of the data.}
}
\description{
This function performs the permuted Brunner-Munzel test.
}
\examples{
Y <-  c(1,2,1,3,1,2,2,4,1,1)
N <-  c(3,4,5,2,3,2,1,5,4,3)

dat <- data.frame(
    value = c(Y, N),
    group = factor(rep(c("Y", "N"), c(length(Y), length(N))),
                   levels = c("Y", "N"))
)

\dontrun{
brunner.munzel.permutation.test(Y, N)
brunner.munzel.permutation.test(value ~ group, data = dat)
}

}
\references{
Karin Neubert and Edgar Brunner, ``A studentized permutation test for the non-parametric Behrens-Fisher problem'', Computational Statistics and Data Analysis, Vol. 51, pp. 5192-5204 (2007).
}
\seealso{
This function is made in reference to following cites (in Japanese): Prof. Haruhiko Okumura \url{https://oku.edu.mie-u.ac.jp/~okumura/stat/brunner-munzel.html} and Anonymous \url{https://blog.goo.ne.jp/r-de-r/e/83dc811baf41ecfe469fa794a4c51b84?fm=rss&utm_medium=twitter&utm_source=twitterfeed}
}
