% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunner.munzel.test.R
\name{brunner.munzel.test}
\alias{brunner.munzel.test}
\alias{brunner.munzel.test.default}
\alias{brunner.munzel.test.formula}
\title{brunner.munzel.test}
\usage{
brunner.munzel.test(x, ...)

\method{brunner.munzel.test}{default}(x, y, alternative = c("two.sided",
  "greater", "less"), alpha = 0.05, ...)

\method{brunner.munzel.test}{formula}(formula, data, subset, na.action,
  ...)
}
\arguments{
\item{x}{the numeric vector of data values from the sample 1.}

\item{\dots}{further arguments to be passed to or from methods
(This argument is for only formula).}

\item{y}{the numeric vector of data values from the sample 2.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. User can specify just the initial letter.}

\item{alpha}{significance level, default is 0.05 for 95\% confidence
interval.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs}
is a numeric variable giving the data values and \code{rhs} a factor
with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to
\code{getOption("na.action")}.}
}
\value{
A list containing the following components:
 \item{statistic}{the Brunner--Munzel test statistic.}
 \item{parameter}{the degrees of freedom.}
 \item{conf.int}{the confidence interval.}
 \item{p.value}{the \eqn{p}-value of the test.}
 \item{data.name}{a character string giving the name of the data.}
 \item{estimate}{an estimate of the effect size, i.e., \eqn{P(X < Y) + 0.5 * P(X =Y )}}
}
\description{
This function performs the Brunner--Munzel test for stochastic
equality of two samples, which is also known as the Generalized Wilcoxon
Test. \code{NA}s from the data are omitted.
This function is a wrapper of lawstat::brunner.munzel.test
 by Vyacheslav Lyubchich et al.
 Please see lawstat package.
}
\note{
There exist discrepancies with Brunner and Munzel (2000)
 because there is a typo in the paper. The corrected version is
 in Neubert and Brunner (2007) (e.g., compare the estimates for
 the case study on pain scores).
The current R function follows Neubert and Brunner (2007).
}
\examples{

## Pain score on the third day after surgery for 14 patients under
## the treatment Y and 11 patients under the treatment N
## (see Brunner and Munzel, 2000; Neubert and Brunner, 2007).

Y <- c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 2, 4, 1, 1)
N <- c(3, 3, 4, 3, 1, 2, 3, 1, 1, 5, 4)

brunner.munzel.test(Y, N)

##       Brunner-Munzel Test
## data: Y and N
## Brunner-Munzel Test Statistic = 3.1375,  df = 17.683, p-value = 0.005786
## 95 percent confidence interval:
##  0.5952169 0.9827052
## sample estimates:
## P(X<Y)+.5*P(X=Y)
##        0.788961


## Formula interface.
dat <- data.frame(
    value = c(Y, N),
    group = rep(c("Y", "N"), c(length(Y), length(N)))
)

brunner.munzel.test(value ~ group, data = dat)

##       Brunner-Munzel Test
## data:  value by group
## Brunner-Munzel Test Statistic = -3.1375, df = 17.683, p-value =
## 0.005786
## 95 percent confidence interval:
##  0.01729479 0.40478314
## sample estimates:
## P(X<Y)+.5*P(X=Y)
##         0.211039

## The same p-value is obtained.

}
