% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocTGH.R
\name{posthocTGH}
\alias{posthocTGH}
\alias{posthocTGH.default}
\alias{posthocTGH.formula}
\title{posthocTGH}
\usage{
posthocTGH(...)

\method{posthocTGH}{default}(y, x, method = c("games-howell", "tukey"),
  conf.level = 0.95, digits = 2, p.adjust = "none",
  formatPvalue = TRUE, ...)

\method{posthocTGH}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{\dots}{further arguments to be passed to or from methods
(this argument is only for formula).}

\item{y}{y has to be a numeric vector.}

\item{x}{x has to be vector that either is a factor or can be converted into
one.}

\item{method}{Which post-hoc tests to conduct. Valid values are "tukey" and
"games-howell".}

\item{conf.level}{Confidence level of the confidence intervals.}

\item{digits}{The number of digits to show in the output.}

\item{p.adjust}{Any valid \code{\link{p.adjust}} method.}

\item{formatPvalue}{Whether to format the p values according to APA
standards (i.e. replace all values lower than .001 with '<.001'). This only
applies to the printing of the object, not to the way the p values are
stored in the object.}

\item{formula}{A formula specifying the model.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the formula
\code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
to be used.}

\item{na.action}{a function which indicates what should happen when the data
contain \code{NA}s. Defaults to \code{getOption("na.action")}.}
}
\value{
A list of three elements: \item{input}{List with input arguments}
\item{intermediate}{List of intermediate objects.} \item{output}{List with
two objects 'tukey' and 'games.howell', containing the outcomes for the
respective post-hoc tests.}
}
\description{
This function is used by the 'oneway' function for oneway analysis of
variance in case a user requests post-hoc tests using the Tukey or
Games-Howell methods.
This function is a wrapper of userfriendlyscience::posthocTGH
 by Gjalt-Jorn Peters (Open University of the Netherlands) & Jeff Bagget
 (University of Wisconsin - La Crosse)
 Please see userfriendlyscience package.
}
\examples{
### Compute post-hoc statistics using the tukey method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet, method="tukey")
posthocTGH(weight ~ Diet, data = ChickWeight, method="tukey")

### Compute post-hoc statistics using the games-howell method
posthocTGH(y=ChickWeight$weight, x=ChickWeight$Diet)
posthocTGH(weight ~ Diet, data = ChickWeight)

}
